consolidate cm.gausswin~
