--- cm.gausswin~/cmwindows.h
@@ -1,6 +1,0 @@
-#ifndef _cmwindows_h
-#define _cmwindows_h
-
-double cm_gauss(long *pos, long *length, double *alpha);
-
-#endif
